adding cross val
